OCR/requirements.txt
pytesseract
pdf2image
docx2pdf
opencv-python
Pillow
filetype